refactor namespace for custom views #2
